Bug 17972 - Reformat acquisitions sidebar menu with acquisitions and administration sections
This patch makes changes to acquisitions-menu.inc so that the style
matches that of the Administration and Tools sidebar menus. The links
are now divided up between Acquisitions (/acqui/) and Administration
(/admin/). The link-highlighting JavaScript used elsewhere has been
added.

To test, apply the patch and view the Acquisitions home page.

- Check that the "Acquisitions home" link is highlighted (bold).
- Test the other Acquisitions links to confirm that highlighting works.
- Click any link in the Administration section of the sidebar menu.
- Confirm that everything looks correct and works correctly.

Signed-off-by: Aleisha Amohia <aleishaamohia@hotmail.com>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>