Bug 18254: Remove call to GetItemsByBiblioitemnumber call from additem.pl
The code from additem to delete all the items of a bibliographic record
is very ackward.
This patch simplifies the algorithm and make the code more readable.

Test plan:
Remove all the items of a bibliographic records
If at least 1 item is checked out you should get an error.

No change with the current behavior is expected.

Signed-off-by: Alex Buckley <alexbuckley@catalyst.net.nz>

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>